Move navigation up one component
